requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CVAT SDK
cvat-sdk==2.9.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Templates
jinja2==3.1.3

# HTTP client
httpx==0.26.0

# Environment
python-dotenv==1.0.0
